Mappings go exlusively one-way only, so there is no point in a common layer.
